Revert "feat(fips): disable kerberos in fips mode"
