+++ gssapi_kerberos.go
@@ -1,5 +1,3 @@
-//go:build !requirefips
-
 package sarama
 
 import (
@@ -20,6 +18,30 @@ import (
 	"github.com/jcmturner/gokrb5/v8/types"
 )
 
+const (
+	TOK_ID_KRB_AP_REQ   = 256
+	GSS_API_GENERIC_TAG = 0x60
+	KRB5_USER_AUTH      = 1
+	KRB5_KEYTAB_AUTH    = 2
+	KRB5_CCACHE_AUTH    = 3
+	GSS_API_INITIAL     = 1
+	GSS_API_VERIFY      = 2
+	GSS_API_FINISH      = 3
+)
+
+type GSSAPIConfig struct {
+	AuthType           int
+	KeyTabPath         string
+	CCachePath         string
+	KerberosConfigPath string
+	ServiceName        string
+	Username           string
+	Password           string
+	Realm              string
+	DisablePAFXFAST    bool
+	BuildSpn           BuildSpnFunc
+}
+
 type GSSAPIKerberosAuth struct {
 	Config                *GSSAPIConfig
 	ticket                messages.Ticket
@@ -35,6 +57,8 @@ type KerberosClient interface {
 	CName() types.PrincipalName
 	Destroy()
 }
+
+type BuildSpnFunc func(serviceName, host string) string
 
 // writePackage appends length in big endian before the payload, and sends it to kafka
 func (krbAuth *GSSAPIKerberosAuth) writePackage(broker *Broker, payload []byte) (int, error) {
