+++ kerberos_client.go
@@ -1,5 +1,3 @@
-//go:build !requirefips
-
 package sarama
 
 import (
